The End

+++ src/limiter/projection.cpp
@@ -82,7 +82,7 @@ void Projection::makeProjection(Array2D &u, const Array2D x, int i, int order)
     for(int j=0;j<u.size()[1];j++){
         u(i,j) = 0.0;
         for(int k = 0;k<order;k++){
-        u(i,j) += pow(x(i,j),k)*coeff_(k,0);
+        u(i,j) += pow(x(i,j),double(k))*coeff_(k,0);
         //std::cout<<" ij"<<i<<" "<<j<<" U "<<u(i,j)<<" pow "<<pow(x(i,j),k)<<" coeff "<< coeff_(k,0)<<std::endl;
         }
     }
@@ -135,7 +135,7 @@ Array2D Projection::MakeMonomBasis(const Array2D &u, int i, int order)
 
     for(int j=0;j<u.size()[1];j++){
         for(int k=0;k<order;k++){
-            solution(j,k) = pow(u(i,j),k);
+            solution(j,k) = pow(u(i,j),double(k));
         }
     }
